# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parent_id:
        type: string
      slug:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    required:
    - name
    type: object
...
  models.CreateReviewDTO:
    properties:
      body:
        type: string
      book_id:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
    required:
    - book_id
    - rating
    type: object
  models.ErrorResponseDTO:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Review:
    properties:
      body:
        type: string
      book_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
...
  models.SuccessResponseDTO:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  models.UpdateReviewDTO:
    properties:
      body:
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
    type: object
...
  models.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      group:
        $ref: '#/definitions/models.UserGroup'
      group_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.UserRole'
      updated_at:
        type: string
    required:
    - email
    type: object
  models.UserGroup:
    properties:
      allowed_categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      can_download:
        type: boolean
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      loan_days:
        type: integer
      max_books:
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/models.UserGroupType'
      updated_at:
        type: string
    required:
    - name
    - type
    type: object
  models.UserGroupType:
    enum:
    - student
    - free
    - subscriber
    type: string
    x-enum-varnames:
    - GroupTypeStudent
    - GroupTypeFree
    - GroupTypeSubscriber
...
  models.UserRole:
    enum:
    - admin
    - librarian
    - reader
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleLibrarian
    - RoleReader
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for a digital library application, providing functionalities for
    book management, user authentication, and reading features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Digital Library API
  version: "2.0"
paths:
...
  /reviews:
    post:
      consumes:
      - application/json
      description: Adds a new review for a book by the authenticated user.
      parameters:
      - description: Review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.CreateReviewDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Create a new review
      tags:
      - Reviews
  /reviews/{id}:
    delete:
      description: Deletes a review by its ID. Can be done by the author or an admin.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - Reviews
    put:
      consumes:
      - application/json
      description: Updates a review written by the authenticated user.
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      - description: Review data to update
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/models.UpdateReviewDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - Reviews
  /reviews/book/{book_id}:
    get:
      description: Retrieves all reviews for a specific book.
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponseDTO'
      summary: Get reviews for a book
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"